% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConvertToCat.R
\name{ConvertToCat}
\alias{ConvertToCat}
\title{Convert Variables to Factor}
\usage{
ConvertToCat(df, nam)
}
\arguments{
\item{df}{A data frame}

\item{nam}{A vector of your categorical variable names}
}
\value{
Variables converted to factors
}
\description{
Convert Variables to Factor
}
